Fix #22: race conditions in workers.
This turned out to be quite a big deal. Redis doesn't have super-easy transactionality. I ended up building a dedicated `ItemListStore` object to be an abstraction over Redis; this made the higer-level code clean again.

As part of this cleanup I also separated the worker process from the web process, so posting a webhook no longer does the transfer to Airtable unless you explicitly ask for it.  So this changelist also fixes #9.