More work to fix #22: race conditions in workers.
The optimistic locking conflicts were not being handled correctly.  In addition, the approach of waiting in the item list processing loop for retries to become available caused tasks to multiply.  Now we use a better approach, which is to publish events both when new items lists arrive and when retry item lists become ready.