More work on #22: support embedding a worker task in the web server.
Since Heroku dev and stage environments are running free dynos, we can't have a separate dyno running the worker process.  But since the worker is itself asynchronous, we can just run it as a task inside the web process (similar to the way the webhook API allows running a worker just long enough to transfer the received item).  So this change cleans up that approach and allows the Heroku environment to control whether an embedded worker is running or not.