Fix #22: race conditions in transfer processing (#23)
* Fix #22: race conditions in workers.

This turned out to be quite a big deal. Redis doesn't have super-easy transactionality. I ended up building a dedicated `ItemListStore` object to be an abstraction over Redis; this made the higer-level code clean again.

As part of this cleanup I also separated the worker process from the web process, so posting a webhook no longer does the transfer to Airtable unless you explicitly ask for it.  So this changelist also fixes #9.

* More work to fix #22: race conditions in workers.

The optimistic locking conflicts were not being handled correctly.  In addition, the approach of waiting in the item list processing loop for retries to become available caused tasks to multiply.  Now we use a better approach, which is to publish events both when new items lists arrive and when retry item lists become ready.

* More work on #22: support embedding a worker task in the web server.

Since Heroku dev and stage environments are running free dynos, we can't have a separate dyno running the worker process.  But since the worker is itself asynchronous, we can just run it as a task inside the web process (similar to the way the webhook API allows running a worker just long enough to transfer the received item).  So this change cleans up that approach and allows the Heroku environment to control whether an embedded worker is running or not.

* Get ready to release v2.1.1.

Aside from a spurious "error while closing socket" message (which only happens in the Heroku environment) things seem ready for release.  I have removed the extraneous "debug=" command to uvicorn (which does nothing) and updated to the latest versions of things.