Fix #57: Action Network phone numbers not being transferred. (#58)
Now they are, along with the mobile opt-out field.

This transformation adds new fields, so it requires an entire re-import of all the existing people.  The current compare-and-update code wouldn't do that because the mod dates hadn't changed, so I added an argument to pass into that code to override the mod date and force the transfer.