Fix #64: watch for duplicate records; don't make new dups.
This was mostly about more careful cache loading and having a way of knowing which existing duplicate records to get rid of in Airtable.

I also put in a number of cache speed-ups, and a lot of guards against creating duplicate records.py

Keeping the entire record set for events and shifts in the redis-cache will definitely speed things up, because we won't have to fetch all the time to check for field updates.